Split up Utils into functions that can be imported directly
