+++ assets/common/utils/index.js
@@ -1,2 +1,4 @@
-export { csrfSafeMethod } from './csrf';
-export default {};
+export * from './alert';
+export * from './dom';
+export * from './string';
+export * from './request';
