+++ assets/common/utils/string.js
@@ -1,0 +1,7 @@
+export const format = (string, ...args) => {
+  string.replace(/{(\d+)}/g, (match, number) => (
+    args[number] !== undefined ? args[number] : match
+  ));
+};
+
+export { format as default };
