+++ assets/core/index.js
@@ -2,7 +2,8 @@ import 'common/less/gallery.less';
 import 'es6-promise/auto';
 import 'whatwg-fetch';
 import moment from 'moment';
-import { initGoogleMaps, initialize, timeOutAlerts } from './init.js';
+import { timeOutAlerts } from 'common/utils/';
+import { initGoogleMaps, initialize } from './init.js';
 import './less/core.less';
